backend_ai/requirements.txt
opencv-python-headless
ultralytics
easyocr
kafka-python
requests
python-dotenv
